feat: support publish npm
