+++ packages/badanmu/.gitignore
@@ -3,3 +3,5 @@ node_modules
 logs
 *.log
 dist
+cjs
+esm
