+++ packages/badanmu/package.json
@@ -2,7 +2,8 @@
   "name": "badanmu",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "cjs/index.js",
+  "module": "esm/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "cli": "ts-node ./src/cli.ts",
@@ -12,7 +13,8 @@
     "start": "cross-env NODE_ENV=production node ./dist/server.js",
     "dev-client": "npx test-client",
     "prod-client": "cross-env NODE_ENV=production npx test-client",
-    "lint": "eslint ./src --ext .js,.ts --fix"
+    "lint": "eslint ./src --ext .js,.ts --fix",
+    "prepublish": "tsc -p tsconfig.cjs.json && tsc -p tsconfig.esm.json"
   },
   "keywords": [],
   "author": "",
@@ -42,5 +44,10 @@
     "prettier": "^2.2.1",
     "test-client": "^1.0.0",
     "ts-node": "^9.1.1"
-  }
+  },
+  "files": [
+    "cjs",
+    "esm",
+    "src"
+  ]
 }
