+++ packages/badanmu/src/index.ts
@@ -1,0 +1,1 @@
+export { default as createClient } from './createClient'
