+++ packages/badanmu/tsconfig.cjs.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "cjs",
+    "declaration": true
+  },
+  "exclude": ["src/**/*.js"]
+}
