+++ packages/badanmu/tsconfig.esm.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "esm",
+    "target": "ESNext",
+    "module": "ESNext",
+    "declaration": true
+  },
+  "exclude": ["src/**/*.js"]
+}
